+++ src/algorithms/dfs.h
@@ -11,7 +11,7 @@
 
 // Check if a node is valid to visit
 bool dfs_is_valid(int grid[HEIGHT][WIDTH], node current, node target);
-int dfs(int grid[HEIGHT][WIDTH],node target,node current,int* tiles,int* total_tiles,bool firstcase);
+int dfs(int grid[HEIGHT][WIDTH],node target,node current,int* tiles,long long* total_tiles,bool firstcase);
